+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `Producer` functions
+    - `produceWithTrace`
+    - `produceSingleWithTrace`
+    - `produceWithTraceTo`
 
 ## 15.1.0 - 2021-07-26
 - Update dependencies